% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{app}
\alias{app}
\alias{new_app}
\title{Create a new web application}
\usage{
new_app()
}
\description{
It's API:
}
\details{
\if{html}{\out{<div class="r">}}\preformatted{app <- new_app()
}\if{html}{\out{</div>}}\if{html}{\out{<div class="r">}}\preformatted{app$get_config(key)
app$set_config(key, value)
app$path()
}\if{html}{\out{</div>}}\if{html}{\out{<div class="r">}}\preformatted{app$all(path, ...)
app$get(path, ...)
app$post(path, ...)
app$use(..., path = "/")
}\if{html}{\out{</div>}}\if{html}{\out{<div class="r">}}\preformatted{app$engine(ext, engine)
}\if{html}{\out{</div>}}\if{html}{\out{<div class="r">}}\preformatted{app$locals
}\if{html}{\out{</div>}}\if{html}{\out{<div class="r">}}\preformatted{app$listen(port = NULL)
}\if{html}{\out{</div>}}
}
