% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{http_time_stamp}
\alias{http_time_stamp}
\title{Format a time stamp for HTTP}
\usage{
http_time_stamp(t = Sys.time())
}
\arguments{
\item{t}{Date-time value to format, defaults to the current date and
time. It must be a POSIXct object.}
}
\value{
Character vector, formatted date-time.
}
\description{
Format a time stamp for HTTP
}
