% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{ids}
\alias{ids}
\title{Generate random ids.}
\usage{
ids(
  data_set = c("cran", "cats", "dogs", "pokemon", "superherores", "supervillains",
    "dinosaurs", "animals"),
  n = 1,
  unique = TRUE,
  join = c("empty", "snake", "dot"),
  case = c("lower", "upper", "title", "keep"),
  pkgname = TRUE
)
}
\arguments{
\item{data_set}{Data set to use. See currently included data sets
below.}

\item{n}{Number of ids to generate.}

\item{unique}{Whether to make sure that ids are unique by appending
numbers to them.}

\item{join}{How to join the words. Only applies if data set has
multiple words.}

\item{case}{Whether to convert words to lower or upper case, or keep
them as they are.}

\item{pkgname}{Whether to make sure that the generated ids are valid
package names, by dropping forbidden characters, and potentially
prefixing with \code{pkg}.}
}
\description{
Generate random ids.
}
\details{
\subsection{Data sets}{
\itemize{
\item \code{cran}: names of CRAN packages.
\item \code{cats}: cat names, from https://github.com/sindresorhus/cat-names
\item \code{dogs}: dog names, from https://github.com/sindresorhus/dog-names
\item \code{pokemon}: Pokemon names, from https://github.com/sindresorhus/pokemon
\item \code{superheroes}: Superhero names, from
https://github.com/sindresorhus/superheroes
\item \code{supervillains}: Supervillain names, from
https://github.com/sindresorhus/supervillains
\item \code{dinosaurs}: dinosaur names, from https://github.com/dariusk/corpora
\item \code{animals}: adjective-animals, from
https://github.com/a-type/adjective-adjective-animal
}
}
}
