% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-app-process.R
\name{local_app_process}
\alias{local_app_process}
\title{TODO}
\usage{
local_app_process(
  app,
  ...,
  .env = parent.frame(),
  .teardown_env = parent.frame()
)
}
\arguments{
\item{.teardown_env}{}
}
\description{
TODO
}
