% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw-multipart.R
\name{mw_multipart}
\alias{mw_multipart}
\title{Parse a multipart HTTP request body}
\usage{
mw_multipart(type = "multipart/form-data")
}
\arguments{
\item{type}{Content type to match before parsing. If it does not
match, then the request object is not modified.}
}
\value{
Handler function.
}
\description{
Adds the parsed form fields in the \code{form} element of the request and
the parsed files to the \code{files} element.
}
\examples{
app <- new_app()
app$use(mw_multipart())
app
}
\seealso{
Other middleware: 
\code{\link{mw_cookie_parser}()},
\code{\link{mw_etag}()},
\code{\link{mw_json}()},
\code{\link{mw_log}()},
\code{\link{mw_raw}()},
\code{\link{mw_static}()},
\code{\link{mw_text}()},
\code{\link{mw_urlencoded}()}
}
\concept{middleware}
