% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw-raw.R
\name{mw_raw}
\alias{mw_raw}
\title{Middleware to read the raw body of a request}
\usage{
mw_raw(type = "application/octet-stream")
}
\arguments{
\item{type}{Content type to match. If it does not match, then the
request object is not modified.}
}
\value{
Handler function.
}
\description{
Adds the raw body, as a raw object to the \code{raw} field of the request.
}
\examples{
app <- new_app()
app$use(mw_raw())
app
}
\seealso{
Other middleware: 
\code{\link{mw_cookie_parser}()},
\code{\link{mw_etag}()},
\code{\link{mw_json}()},
\code{\link{mw_log}()},
\code{\link{mw_multipart}()},
\code{\link{mw_static}()},
\code{\link{mw_text}()},
\code{\link{mw_urlencoded}()}
}
\concept{middleware}
