% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth.R
\name{oauth2_resource_app}
\alias{oauth2_resource_app}
\alias{oauth2_third_party_app}
\alias{oauth2_login}
\title{OAuth 2.0}
\usage{
oauth2_resource_app(
  access_duration = 3600L,
  refresh_duration = 7200L,
  refresh = TRUE,
  seed = NULL
)

oauth2_third_party_app(name = "Third-Party app")

oauth2_login(login_url)
}
\arguments{
\item{access_duration}{After how many seconds should access tokens expire.}

\item{refresh_duration}{After how many seconds should refresh tokens expire
(ignored if \code{refresh} is \code{FALSE}).}

\item{refresh}{Should a refresh token be returned (logical).}

\item{seed}{Random seed used when creating tokens.}

\item{name}{Name of the third-party app}

\item{login_url}{The login URL of the third party app.}
}
\value{
a \code{webfakes} app

A named list with
\itemize{
\item \code{login_response} The curl HTTP response object for the login
page.
\item \code{token_response} The curl HTTP response object for submitting
the login page.
}
}
\description{
The webfakes package comes with two fake apps that allow to imitate the
OAuth2.0 flow in your test cases. (See \href{https://aaronparecki.com/oauth-2-simplified/}{Aaron Parecki’s tutorial} for a good
introduction to OAuth2.0.) One app (\code{oauth2_resource_app()}) is the API
server that serves both as the resource and provides authorization.
\code{oauth2_third_party_app()} plays the role of the third-party app. They
are useful when testing or demonstrating code handling OAuth2.0
authorization, token caching, etc. in a package. The apps can be used in
your tests directly, or you could adapt one or both of them to better
mimic a particular OAuth2.0 flow.

It works with \code{\link[=oauth2_resource_app]{oauth2_resource_app()}}, and any third party app,
including the fake \code{\link[=oauth2_third_party_app]{oauth2_third_party_app()}}.
}
\details{
For more details see \code{vignette("oauth", package = "webfakes")}.

See \code{test-oauth.R} in webfakes for an example.
}
\section{\code{oauth2_resource_app()}}{

App representing the API server (resource/authorization)
}

\section{\code{oauth2_third_party_app()}}{

App representing the third-party app
}

\concept{OAuth2.0 functions}
