% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{presser_response}
\alias{presser_response}
\title{A presser response object}
\description{
presser creates a \code{presser_response} object for every incoming HTTP
request. This object is passed to every matched route and middleware,
until the HTTP response is sent. It has reference semantics, so handlers
can modify it.
}
\details{
Fields and methods:
\itemize{
\item \code{app}: The \code{presser_app} object itself.
\item \code{locals}: Local variables, the are shared between the handler
functions. This is for the end user.
\item \code{get_header(field)}: Query the currently set response headers. If
\code{field} is not present it return \code{NULL}.
\item \code{on_response(fun)}: Run the \code{fun} handler function just before the
response is sent out. At this point the headers and the body are
already properly set.
\item \code{redirect(path, status = 302)}: Send a redirect response. It sets
the \code{Location} header, and also sends a \code{text/plain} body.
\item \code{render(view, locals = list())}: Render a template page. Searches
for the \code{view} template page, using all registered engine extensions,
and calls the first matching template engine. Returns the filled
template.
\item \code{send_json(object = NULL, text = NULL, ...)}: Send a JSON response.
Either \code{object} or \code{text} must be given. \code{object} will be converted
to JSON using \code{\link[jsonlite:toJSON]{jsonlite::toJSON()}}. \code{...} are passed to
\code{\link[jsonlite:toJSON]{jsonlite::toJSON()}}. It sets the content type appropriately.
\item \code{send(body)}. Send the specified body. \code{body} can be a raw vector,
or HTML or other text. For raw vectors it sets the content type to
\code{application/octet-stream}.
\item \code{send_file(path, root = ".")}: Send a file. Set \code{root = "/"} for
absolute file names. It sets the content type automatically, based
on the extension of the file, if it is not set already.
\item \code{send_status(status)}: Send the specified HTTP status code, without
a response body.
\item \code{set_header(field, value)}: Set a response header.
\item \code{set_status(status)}: Set the response status code.
\item \code{set_type(type)}: Set the response content type. If it contains a \code{/}
character then it is set as is, otherwise it is assumed to be a file
extension, and the corresponding MIME type is set.
}
}
\seealso{
\link{presser_request} for the presser request object.
}
