% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server_opts}
\alias{server_opts}
\title{Presser web server options}
\usage{
server_opts(port = NULL, num_threads = 1)
}
\arguments{
\item{port}{Port to start the web server on. Defaults to a randomly
chosen port.}

\item{num_threads}{Number of request handler threads to use. Typically
you don't need more than one thread, unless you run test cases in
parallel or you make concurrent HTTP requests.}
}
\description{
Presser web server options
}
