% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webfakes-package.R
\docType{package}
\name{webfakes-package}
\alias{webfakes}
\alias{webfakes-package}
\title{webfakes: Fake Web Apps for HTTP Testing}
\description{
Create a web app that makes it easier to test web clients without using the internet. It includes a web app framework with path matching, parameters and templates. Can parse various 'HTTP' request bodies. Can send 'JSON' data or files from the disk. Includes a web app that implements the \url{https://httpbin.org} web service.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://webfakes.r-lib.org/}
  \item \url{https://github.com/r-lib/webfakes#readme}
  \item Report bugs at \url{https://github.com/r-lib/webfakes/issues}
}

}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com}

Other contributors:
\itemize{
  \item  RStudio, Pbc. [copyright holder, funder]
  \item  Civetweb contributors (see inst/credits/ciwetweb.md) [contributor]
  \item  Redoc contributors (see inst/credits/redoc.md) [contributor]
  \item L. Peter Deutsch (src/md5.h) [contributor]
  \item Martin Purschke (src/md5.h) [contributor]
  \item  Aladdin Enterprises (src/md5.h) [copyright holder]
  \item Maëlle Salmon \email{maelle.salmon@yahoo.se} (\href{https://orcid.org/0000-0002-2815-0399}{ORCID}) [contributor]
}

}
\keyword{internal}
