% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpbin.R
\name{httpbin_app}
\alias{httpbin_app}
\title{Generic web app for testing HTTP clients, like httpbin.org}
\usage{
httpbin_app(log = interactive())
}
\arguments{
\item{log}{Whether to log requests to the standard output.}
}
\value{
A \code{presser_app}.
}
\description{
Generic web app for testing HTTP clients, like httpbin.org
}
\examples{
app <- httpbin_app()
proc <- new_app_process(app)
url <- proc$get_url("/get")
resp <- curl::curl_fetch_memory(url)
curl::parse_headers_list(resp$headers)
cat(rawToChar(resp$content))
proc$stop()
}
