% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-app-process.R
\name{local_app_process}
\alias{local_app_process}
\title{App process that is cleaned up automatically}
\usage{
local_app_process(
  app,
  ...,
  .env = parent.frame(),
  .local_envir = parent.frame()
)
}
\arguments{
\item{app}{\code{presser_app} object, the web app to run.}

\item{...}{Passed to \code{\link[=new_app_process]{new_app_process()}}.}

\item{.env}{The environment to run \code{\link[=new_app_process]{new_app_process()}} in.}

\item{.local_envir}{The environment to attach the process cleanup to.
Typically a frame. When this frame finishes, the process is stopped.}
}
\description{
You can start the process with an explicit \verb{$start()} call.
Alternatively it starts up at the first \verb{$url()} or \verb{$get_port()}
call.
}
\seealso{
\code{\link[=new_app_process]{new_app_process()}} for more details.
}
