% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw-cookie-parser.R
\name{mw_cookie_parser}
\alias{mw_cookie_parser}
\title{Middleware to parse Cookies}
\usage{
mw_cookie_parser()
}
\value{
Handler function.
}
\description{
Adds the cookies as the \code{cookies} element of the request object.
}
\details{
It ignores cookies in an invalid format. It ignores duplicate cookies:
if two cookies have the same name, only the first one is included.
}
\seealso{
Other middleware: 
\code{\link{mw_etag}()},
\code{\link{mw_json}()},
\code{\link{mw_log}()},
\code{\link{mw_multipart}()},
\code{\link{mw_raw}()},
\code{\link{mw_static}()},
\code{\link{mw_text}()},
\code{\link{mw_urlencoded}()}
}
\concept{middleware}
