% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mw-etag.R
\name{mw_etag}
\alias{mw_etag}
\title{Middleware that add an \code{Etag} header to the response}
\usage{
mw_etag(algorithm = "crc32")
}
\arguments{
\item{algorithm}{Checksum algorithm to use. See the \code{also} argument of
\code{\link[digest:digest]{digest::digest()}} for the possible values.}
}
\value{
Handler function.
}
\description{
It needs the digest package to calculate the value of the header.
}
\seealso{
Other middleware: 
\code{\link{mw_json}()},
\code{\link{mw_log}()},
\code{\link{mw_multipart}()},
\code{\link{mw_raw}()},
\code{\link{mw_static}()},
\code{\link{mw_text}()},
\code{\link{mw_urlencoded}()}
}
\concept{middleware}
