% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauth.R
\name{oauth2.0}
\alias{oauth2_third_party_app}
\title{OAuth 2.0}
\usage{
oauth2_resource_app()

oauth2_third_party_app(name = "Third-Party app")
}
\value{
a \code{webfakes} app
}
\description{
These two apps allow to imitate the simplified OAuth2.0 flow described
in \href{https://aaronparecki.com/oauth-2-simplified/}{Aaron Parecki’s tutorial}. One app
(\code{oauth2_resource_app()}) is the API server (resource/authorization),
the other plays the role of a third-party app. They are useful when
testing or demonstrating code handling OAuth2.0 authorization, token
caching, etc. in a package: The apps can be used in your tests directly,
or you could adapt one of them / both of them to better mimick a
particular OAuth2.0 flow.
}
\details{
For more details see \code{vignette("oauth", package = "webfakes")}.
}
\section{\code{oauth2_resource_app()}}{

App representing the API server (resource/authorization)
}

\section{\code{oauth2_third_party_app()}}{

App representing the third-party app
}

